FIX: code structure
