+++ network/include/protocol/Protocol.hpp
@@ -7,19 +7,20 @@
 
 #pragma once
 
-#include "Room.hpp"
-#include "SmartBuffer.hpp"
+#include <SmartBuffer.hpp>
 #include <memory>
 #include <unordered_map>
 
 class Protocol {
   public:
     enum OpCode {
-        DEFAULT = 0,
-        CREATE_ROOM = 1,
-        ROOM_INFO = 2,
-        JOIN_ROOM = 3,
-        DELETE_ROOM = 4
+        DEFAULT,
+        CREATE_ROOM,
+        CREATE_ROOM_CALLBACK,
+        JOIN_ROOM,
+        JOIN_ROOM_CALLBACK,
+        DELETE_ROOM,
+        DELETE_ROOM_CALLBACK,
     };
 
     Protocol();
@@ -30,8 +31,4 @@ class Protocol {
     void createRoom(int clientSocket, SmartBuffer& smartBuffer);
     void joinRoom(int clientSocket, SmartBuffer& smartBuffer);
     void deleteRoom(int clientSocket, SmartBuffer& smartBuffer);
-    void sendRoomInfo(int clientSocket, int32_t roomCode);
-
-    std::unordered_map<int32_t, std::shared_ptr<Room>> rooms;
-    std::mutex roomMutex;
 };
