+++ network/src/util/Logger.cpp
@@ -5,7 +5,7 @@
 ** Logger.cpp
 */
 
-#include "Logger.hpp"
+#include "util/Logger.hpp"
 #include <chrono>
 #include <ctime>
 #include <iomanip>
